Separate out processing_lambda tests
